Commit.
Implementing jQuery method fadeOut in all displayed messages after
click button in all functions: both authentication related and
CRUD action related. The intention is to improve the UI | UX experience
of the web application.

Uncommenting some other console.log messages, so it won't be displayed
in the console inspector of the client when the web application will be
running.