Commit.
Modifications on file ui.js from both auth and songs directories. File
index.html edited as well.

In order to provide improvements in the UI | UX experience, messages
for each action of the application is rendered right above the click
buttons of each functionality, instead of on the top right bellow the
header for all actions.